@@ -1,21 +1,22 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
+    <TargetFrameworks>net6.0</TargetFrameworks>
     <ServerGarbageCollection>true</ServerGarbageCollection>
     <ConcurrentGarbageCollection>true</ConcurrentGarbageCollection>
     <ThreadPoolMinThreads>500</ThreadPoolMinThreads>
   </PropertyGroup>
 
  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="3.1">
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="6.0">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="3.1" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1" />
-    <PackageReference Include="MySqlConnector" Version="2.1.9" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.0" />
-    <PackageReference Include="HtmlSanitizer" Version="6.0.453" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="6.0"/>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="6.0"/>
+    <PackageReference Include="MySqlConnector" Version="2.1.10"/>
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.3.1"/>
+    <PackageReference Include="HtmlSanitizer" Version="6.0.453"/>
   </ItemGroup>
 </Project>
