@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using BossFight.Extentions;
 using MySqlConnector;
-using Newtonsoft.Json;
+using System.Text.Json.Serialization;
 
 namespace BossFight.Models.DB
 {
